all kinds of updates

+++ LEDUtils.cpp
@@ -1,9 +1,10 @@
-void led_blink(byte LEDpin, byte onMS, byte offMS, byte loops) {
-  for (byte i=0; i<loops; i++)  {
+#include <Esp.h>
+
+void led_blink(int LEDpin, int onMS, int offMS, int loops) {
+  for (uint8_t i=0; i<loops; i++)  {
     digitalWrite(LEDpin, HIGH);
     delay(onMS);
     digitalWrite(LEDpin, LOW);
     delay(offMS);
   }
 }
-
